+++ gendarme/FSharpExamples/ConsiderConvertingMethodToPropertyRule.fs
@@ -1,0 +1,8 @@
+﻿namespace ConsiderConvertingMethodToProperty
+
+module Instance =
+  let mutable isRunner = false
+  let mutable isOpenCover = false
+
+  let isOpenCoverRunner() =
+    isOpenCover && isRunner